+++ src/components/maps/CloseButton.tsx
